App component test added
