+++ src/tests/components/App.test.js
@@ -1,0 +1,23 @@
+import React from 'react';
+import { render, screen } from '@testing-library/react';
+import renderer from 'react-test-renderer';
+import { BrowserRouter } from 'react-router-dom';
+import App from '../../components/App';
+
+describe('App', () => {
+  test('renders App component ', () => {
+    render(<App />);
+    // eslint-disable-next-line no-unused-expressions
+    expect(screen.getByText('Let’s do some math!')).toBeInTheDocument;
+  });
+  test('should match with snapshot', () => {
+    const tree = renderer
+      .create(
+        <BrowserRouter>
+          <App />
+        </BrowserRouter>,
+      )
+      .toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+});
